requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0
werkzeug==3.0.1
Flask-Mail==0.9.1
